+++ app/Http/Controllers/Auth/LoginController.php
@@ -3,10 +3,9 @@
 namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
-use App\User;
+use App\Model\User;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 
 class LoginController extends Controller
 {
@@ -50,7 +49,7 @@ class LoginController extends Controller
     protected function authenticated(Request $request, $user)
     {
         if ($user && $user->isAdmin()) {
-            return redirect()->route('admin.index')->with('success', 'Vous êtes connecté maître.');
+            return redirect()->route('home.index')->with('success', 'Vous êtes connecté maître.');
         }
         return redirect()->route('home.index')->with('success', 'Vous êtes connecté.');
     }
