+++ app/Http/Controllers/PostsController.php
@@ -1,8 +1,8 @@
 <?php
 namespace App\Http\Controllers;
 
-use App\Category;
-use App\Post;
+use App\Model\Category;
+use App\Model\Post;
 use App\Repository\CategoryRepository;
 use App\Repository\PostRepository;
 use Illuminate\Auth\AuthManager;
