+++ app/Model/Post.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace App;
+namespace App\Model;
 
 use App\Favorite\Favorite;
-use App\Favorite\HasFavorites;
 use App\Traits\HasSlug;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
