+++ database/factories/CategoryFactory.php
@@ -13,7 +13,7 @@ use Faker\Generator as Faker;
 |
 */
 
-$factory->define(App\Category::class, function (Faker $faker) {
+$factory->define(App\Model\Category::class, function (Faker $faker) {
 
     return [
         'name' => $faker->name,
