+++ database/seeds/PostsTableSeeder.php
@@ -14,6 +14,6 @@ class PostsTableSeeder extends Seeder
      */
     public function run()
     {
-        factory(\App\Post::class, 10)->create();
+        factory(\App\Model\Post::class, 10)->create();
     }
 }
