+++ database/seeds/UsersTableSeeder.php
@@ -11,6 +11,6 @@ class UsersTableSeeder extends Seeder
      */
     public function run()
     {
-        factory(\App\User::class, 5)->create();
+        factory(\App\Model\User::class, 5)->create();
     }
 }
